refactor: 200줄 넘긴 파일 둘 분리 #67
CLAUDE.md §폴더·네이밍(200줄↑ 분리)을 넘긴 것 중 이번에 만든 둘.

- subscribe-card(267) → checkout-panel 분리. "무엇이 들어 있나"(왼쪽)와
  "얼마를 내나"(오른쪽)는 성격이 다른 칸인데 한 덩어리로 붙어 있었다
- role-screens(238) → role-mocks 분리. 넷을 한 뼈대로 묶고 나니 파일의 절반이
  데이터였다 — 문구를 고치러 들어가서 마크업을 건드리는 일이 없게
